+++ ft_strmapi.c
@@ -16,7 +16,7 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	unsigned int	i;
+	unsigned long	i;
 	char			*str;
 
 	i = 0;
